requirements.txt
absl-py==2.4.0
annotated-types==0.7.0
anyio==4.13.0
av==17.0.0
cerebras==0.0.1
cerebras_cloud_sdk==1.67.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
cryptography==46.0.7
cuda-bindings==12.9.6
cuda-pathfinder==1.4.4
cuda-toolkit==12.6.3
cycler==0.12.1
distro==1.9.0
fastapi==0.118.2
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
fvcore==0.1.5.post20221221
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iopath==0.1.10
Jinja2==3.1.6
kiwisolver==1.5.0
lap==0.5.13
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.33
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.85
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.6.77
# "opencv-contrib-python" includes "cv2" (do not add "opencv-python" — they conflict on install).
opencv-contrib-python==4.13.0.92
packaging==26.0
parameterized==0.9.0
pillow==12.1.1
polars==1.39.3
polars-runtime-32==1.39.3
portalocker==3.2.0
psutil==7.2.2
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.3.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorchvideo==0.1.5
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.14.0
tabulate==0.10.0
termcolor==3.3.0
torch==2.11.0+cu126
torchreid==0.2.5
torchvision==0.26.0+cu126
tqdm==4.67.3
transformers==4.44.2
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.26
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.37.0
yacs==0.1.8
